It is possible to click outside the goto definition input box and
the quick panel stays visible, so this won't work in that case, and
on_close isn't fired when the quick panel really is closed. If you
just care about when the user is typing in this view, it should be
enough to ignore the quick panel view in your plugin because the id
doesn't seem to get reused.

But, I'm not sure how to tell when it  has been closed, without
making changes to Packages/Default/symbols.py which I wouldn't
recommend